+++ Model/ResourceModel/Cron/CronRunLog/Collection.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Ordo\Automation\Model\ResourceModel\Cron\CronRunLog;
+
+use Magento\Framework\Model\ResourceModel\Db\Collection\AbstractCollection;
+use Ordo\Automation\Model\Cron\CronRunLog as CronRunLogModel;
+use Ordo\Automation\Model\ResourceModel\Cron\CronRunLog as CronRunLogResource;
+
+class Collection extends AbstractCollection
+{
+    protected function _construct(): void
+    {
+        $this->_init(CronRunLogModel::class, CronRunLogResource::class);
+    }
+}